move hosted, proxy and group classes into separate packages
