+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/group/ComposerGroupMergingHandler.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.composer.internal;
+package org.sonatype.nexus.repository.composer.internal.group;
 
 import java.util.List;
 import java.util.Map;
