+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/hosted/ComposerHostedUploadHandler.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.composer.internal;
+package org.sonatype.nexus.repository.composer.internal.hosted;
 
 import com.google.common.base.Preconditions;
 import com.google.common.io.ByteStreams;
@@ -22,6 +22,7 @@ import org.sonatype.nexus.repository.composer.AssetKind;
 import org.sonatype.nexus.repository.composer.ComposerContentFacet;
 import org.sonatype.nexus.repository.composer.ComposerFormat;
 import org.sonatype.nexus.repository.composer.ComposerHostedFacet;
+import org.sonatype.nexus.repository.composer.internal.ComposerContentFacetImpl;
 import org.sonatype.nexus.repository.content.fluent.FluentAsset;
 import org.sonatype.nexus.repository.http.HttpResponses;
 import org.sonatype.nexus.repository.importtask.ImportFileConfiguration;
